pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dcypher"
version = "0.1.0"
description = "dcypher"
requires-python = ">=3.12"
dependencies = [
    "uv",
    "fastapi",
    "uvicorn",
    "pytest",
    "pytest-cov",
    "pytest-xdist",
    "pytest-asyncio",
    "requests",
    "aiohttp",
    "ruff",
    "pre-commit",
    "pandas",
    "openfhe",
    "ecdsa>=0.19.1",
    "python-multipart>=0.0.20",
    "bip-utils>=2.9.3",
    "cryptography>=45.0.4",
    "psutil>=7.0.0",
    "textual[syntax]>=3.5.0",
    "liboqs-python>=0.13.0",
]

[project.scripts]
dcypher = "dcypher.cli.main:cli"

[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]

[tool.setuptools.package-dir]
"" = "src"

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "I"]
ignore-init-module-imports = true

[tool.ruff.lint]
# Add this line to ignore the F401 error for specific unused imports
ignore = ["F401"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "E402"]
"src/dcypher/dcypher_front/src/App.test.js" = ["E501"]
"src/dcypher/dcypher_front/src/reportWebVitals.js" = ["E501"]
"src/dcypher/dcypher_front/src/index.js" = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "lf"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
norecursedirs = ["vendor", "openfhe-python"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = [
    "--import-mode=importlib",
    "--ignore=setup.py",
    "--ignore=src/dcypher/dcypher_front",
    "--ignore=src/dcypher/dcypher_back",
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "--tb=short",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=50",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "tui: marks tests as TUI tests",
    "crypto: Crypto context tests that need sequential execution",
    "live_server: Tests that require a live server",
    "e2e: marks tests as end-to-end workflow tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    # Ignore deprecation warnings from websockets library in FastAPI/Starlette
    "ignore:websockets.server.WebSocketServerProtocol is deprecated.*:DeprecationWarning",
    "ignore::pytest.PytestReturnNotNoneWarning",
    "ignore::DeprecationWarning:openfhe.*",
    "ignore::DeprecationWarning:tests.*",
]

[tool.uv.sources]
dcypher-backend = { path = ".", editable = true }
openfhe-python = { path = "./openfhe-python" }
openfhe = { path = "vendor/openfhe-python", editable = true }

[dependency-groups]
dev = ["httpx>=0.28.1", "setuptools>=80.9.0", "textual-dev>=1.7.0"]
